+++ src/main/java/umc/cockple/demo/domain/exercise/service/ExerciseParticipationService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.cockple.demo.domain.exercise.converter.ExerciseConverter;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
+import umc.cockple.demo.domain.exercise.dto.ExerciseCancelDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseJoinDTO;
 import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
 import umc.cockple.demo.domain.exercise.exception.ExerciseException;
@@ -54,9 +55,41 @@ public class ExerciseParticipationService {
         return exerciseConverter.toJoinResponse(savedMemberExercise, exercise);
     }
 
+
+    public ExerciseCancelDTO.Response cancelParticipation(Long exerciseId, Long memberId) {
+
+        log.info("운동 참여 취소 시작 - exerciseId: {}, memberId: {}", exerciseId, memberId);
+
+        Exercise exercise = findExerciseOrThrow(exerciseId);
+        Member member = findMemberOrThrow(memberId);
+        MemberExercise memberExercise = findMemberExerciseOrThrow(exercise, member);
+        exerciseValidator.validateCancelParticipation(exercise);
+
+        exercise.removeParticipation(memberExercise);
+        member.removeParticipation(memberExercise);
+
+        memberExerciseRepository.delete(memberExercise);
+
+        exerciseRepository.save(exercise);
+
+        log.info("운동 참여 취소 완료 - exerciseId: {}, memberId: {}, 현재 참여자 수: {}",
+                exerciseId, memberId, exercise.getNowCapacity());
+
+        return exerciseConverter.toCancelResponse(exercise, member);
+    }
+
+    // ========== 비즈니스 메서드 ============
+
     private boolean isPartyMember(Exercise exercise, Member member) {
         Party party = exercise.getParty();
         return memberPartyRepository.existsByPartyAndMember(party, member);
+    }
+
+    // ========== 조회 메서드 ============
+
+    private Exercise findExerciseOrThrow(Long exerciseId) {
+        return exerciseRepository.findById(exerciseId)
+                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.EXERCISE_NOT_FOUND));
     }
 
     private Exercise findExerciseWithPartyLevelOrThrow(Long exerciseId) {
@@ -69,5 +102,8 @@ public class ExerciseParticipationService {
                 .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.MEMBER_NOT_FOUND));
     }
 
-
+    private MemberExercise findMemberExerciseOrThrow(Exercise exercise, Member member) {
+        return memberExerciseRepository.findByExerciseAndMember(exercise, member)
+                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.MEMBER_EXERCISE_NOT_FOUND));
+    }
 }
